Greatly optimize catalog processing; add batch optimizations to Postgres

+++ src/main/java/io/salad109/conjunctionapi/ingestion/internal/IngestionLogRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IngestionLogRepository extends JpaRepository<IngestionLog, Long> {
+public interface IngestionLogRepository extends JpaRepository<IngestionLog, Integer> {
 }
